problem of pytest solved
